assignment_1/prediction_api/requirements.txt
flask
joblib
pandas
numpy==1.26.4
scikit-learn==1.3.2
six
google-cloud-storage
setuptools
